D3ASIM-3356: Add abstract matching algorithm class | refactor pay_as_bid_matcher
