+++ d3a_interface/utils.py
@@ -30,10 +30,10 @@ from statistics import mean
 from pkgutil import walk_packages
 from redis.exceptions import ConnectionError
 from pendulum import DateTime, from_format, from_timestamp, duration, today, datetime
-from d3a_interface.constants_limits import DATE_TIME_UI_FORMAT, DATE_TIME_FORMAT, TIME_FORMAT, \
-    DATE_TIME_FORMAT_SECONDS, DEFAULT_PRECISION, GlobalConfig, TIME_ZONE, \
-    CN_PROFILE_EXPANSION_DAYS, FLOATING_POINT_TOLERANCE
-from d3a_interface.dataclasses import BidOfferMatch
+from d3a_interface.constants_limits import (
+    DATE_TIME_UI_FORMAT, DATE_TIME_FORMAT, TIME_FORMAT,
+    DATE_TIME_FORMAT_SECONDS, DEFAULT_PRECISION, GlobalConfig, TIME_ZONE,
+    CN_PROFILE_EXPANSION_DAYS)
 
 
 def convert_datetime_to_str_in_list(in_list: List, ui_format: bool = False):
@@ -475,41 +475,3 @@ def sort_list_of_dicts_by_attribute(input_list: List[Dict],
     return sorted(input_list,
                   key=lambda obj: obj.get(attribute),
                   reverse=reverse_order)
-
-
-def perform_pay_as_bid_match(market_offers_bids_list_mapping):
-    """Performs pay as bid matching algorithm.
-
-    There are 2 simplistic approaches to the problem
-        1. Match the cheapest offer with the most expensive bid. This will favor the sellers
-        2. Match the cheapest offer with the cheapest bid. This will favor the buyers,
-           since the most affordable offers will be allocated for the most aggressive buyers.
-    Args:
-        market_offers_bids_list_mapping: dict {market_uuid: {"offers": [...], "bids": [...]}, }
-
-    Returns: List[BidOfferMatch.serializable_dict()]
-    """
-    bid_offer_pairs = []
-    for market_id, data in market_offers_bids_list_mapping.items():
-        bids = data.get("bids")
-        offers = data.get("offers")
-        # Sorted bids in descending order
-        sorted_bids = sort_list_of_dicts_by_attribute(bids, "energy_rate", True)
-        # Sorted offers in descending order
-        sorted_offers = sort_list_of_dicts_by_attribute(offers, "energy_rate", True)
-        already_selected_bids = set()
-        for offer in sorted_offers:
-            for bid in sorted_bids:
-                if bid.get("id") in already_selected_bids or\
-                        offer.get("seller") == bid.get("buyer"):
-                    continue
-                if (offer.get("energy_rate") - bid.get("energy_rate")) <= FLOATING_POINT_TOLERANCE:
-                    already_selected_bids.add(bid.get("id"))
-                    selected_energy = min(bid.get("energy"), offer.get("energy"))
-                    bid_offer_pairs.append(BidOfferMatch(market_id=market_id,
-                                                         bid=bid, offer=offer,
-                                                         selected_energy=selected_energy,
-                                                         trade_rate=bid.get("energy_rate"))
-                                           .serializable_dict())
-                    break
-    return bid_offer_pairs
